requirements.txt
dash
dash-bootstrap-components
pandas
numpy
matplotlib
seaborn
scipy
xgboost
scikit-learn
plotly
gunicorn
statsmodels
psycopg2-binary
sqlalchemy